make go fast YAY
