+++ TeamCode/src/main/java/org/nknsd/teamcode/components/handlers/artifact/MicrowaveScoopHandler.java
@@ -19,7 +19,7 @@ public class MicrowaveScoopHandler implements NKNComponent {
 
     // this state makes the scoop go up and down
     class ScoopActionState extends StateMachine.State {
-        final double SCOOPACTIONTIMEMS = 300;
+        final double SCOOPACTIONTIMEMS = 250;
 
         private boolean scoopResting = false;
